# grpc/profile.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "profile.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/delete/{id}": {
      "delete": {
        "operationId": "DeleteById",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcProfile"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    },
...
  },
  "definitions": {
    "grpcProfile": {
      "type": "object",
      "properties": {
        "firebaseId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "alias": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
...
  }
}